requirements.txt
# Core dependencies
aiohttp>=3.9.0
websockets==12.0
python-dotenv>=1.0.0

# API Integrations
neoapi-sdk==0.1.3
requests>=2.31.0

# Cache & Storage
redis==5.0.1
aioredis==2.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-aiohttp>=1.0.4
aioresponses>=0.7.4
pytest-mock>=3.10.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0 